# backend/docs/swagger.yaml
components:
  schemas:
    api.HealthResponse:
      properties:
        status:
          example: healthy
          type: string
      type: object
    data:
      properties:
        data:
          $ref: '#/components/schemas/api.HealthResponse'
      type: object
    handler.ConnectionRequest:
      properties:
        database_name:
          example: app
          type: string
        driver:
          enum:
          - postgres
          - mysql
          - sqlite
          example: postgres
          type: string
        host:
          example: 127.0.0.1
          type: string
        name:
          example: Local PG
          type: string
        password:
          example: secret
          type: string
        port:
          example: 5432
          type: integer
        ssl_mode:
          enum:
          - disable
          - allow
          - prefer
          - require
          - verify-ca
          - verify-full
          example: disable
          type: string
        username:
          example: appuser
          type: string
      required:
      - database_name
      - driver
      - name
      type: object
...
    handler.ConnectionTestRequest:
      properties:
        database_name:
          example: app
          type: string
        driver:
          enum:
          - postgres
          - mysql
          - sqlite
          example: postgres
          type: string
        host:
          example: 127.0.0.1
          type: string
        password:
          example: secret
          type: string
        port:
          example: 5432
          type: integer
        ssl_mode:
          enum:
          - disable
          - allow
          - prefer
          - require
          - verify-ca
          - verify-full
          example: disable
          type: string
        username:
          example: appuser
          type: string
      required:
      - database_name
      - driver
      type: object
...
    response.APIError:
      properties:
        code:
          type: string
        message:
          type: string
        position:
          type: integer
      type: object
...
    response.ErrorEnvelope:
      properties:
        data:
          description: always null
        error:
          $ref: '#/components/schemas/response.APIError'
        meta:
          additionalProperties: {}
          type: object
        success:
          example: false
          type: boolean
      type: object
...
info:
  description: Local-first database GUI backend. All responses use the standard envelope
    (success, data, error, meta).
  title: DataDeck API
  version: "1.0"
openapi: 3.1.0
paths:
  /connections:
    get:
      description: Returns all persisted connection profiles with credentials omitted.
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Internal Server Error
      summary: List connection profiles
      tags:
      - connections
    post:
      description: Validates and persists a profile. The password is encrypted at
        rest and never returned.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ConnectionRequest'
              description: Connection profile
              summary: request
        description: Connection profile
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Internal Server Error
      summary: Create a connection profile
      tags:
      - connections
  /connections/{id}:
    delete:
      description: Closes any active pool and removes the profile.
      parameters:
      - description: Connection id
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Internal Server Error
      summary: Delete a connection profile
      tags:
      - connections
...
  /connections/test:
    post:
      description: Validates parameters and attempts a temporary connection. Nothing
        is persisted.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handler.ConnectionTestRequest'
              description: Connection parameters
              summary: request
        description: Connection parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  data: {}
                  error:
                    $ref: '#/components/schemas/response.APIError'
                  meta:
                    additionalProperties: {}
                    type: object
                  success:
                    type: boolean
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Request
        "502":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.ErrorEnvelope'
          description: Bad Gateway
      summary: Test connection parameters
      tags:
      - connections
...
servers:
- url: /api/v1